Added a Special Social Media Handler
Because we all know that's at least some of what you were going to do anyways. :p
Most people will use this proxy to access social media and youtube, so it is logical to cache those sites as most people will be happy that they can load these sites quickly.